codereview cleanup
