+++ lib/services/OrderDispatcher.ts
@@ -1,10 +1,8 @@
 import { Logger } from '@aws-lambda-powertools/logger'
 import { OrderType } from '@uniswap/uniswapx-sdk'
 import { NoHandlerConfiguredError } from '../errors/NoHandlerConfiguredError'
-import { DutchV1Order } from '../models/DutchV1Order'
-import { DutchV2Order } from '../models/DutchV2Order'
-import { LimitOrder } from '../models/LimitOrder'
 import { Order } from '../models/Order'
+import { UniswapXOrder } from '../models/UniswapXOrder'
 import { UniswapXOrderService } from './UniswapXOrderService'
 
 export class OrderDispatcher {
@@ -20,9 +18,7 @@ export class OrderDispatcher {
     throw new NoHandlerConfiguredError(order.orderType)
   }
 
-  private isUniswapXOrder(order: Order): order is DutchV1Order | LimitOrder | DutchV2Order {
-    // Once the UniswapXService supports Xv2, add the type check here and change the
-    // typeguard to order is UniswapXOrder
+  private isUniswapXOrder(order: Order): order is UniswapXOrder {
     return (
       order.orderType === OrderType.Dutch ||
       order.orderType === OrderType.Limit ||
